% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_common_words}
\alias{rm_common_words}
\title{Removes common words from corpus}
\usage{
rm_common_words(corpus, common_word_limit, tokenized_corpus = NULL)
}
\arguments{
\item{corpus}{A dataframe of the corpus}

\item{common_word_limit}{An integer defining the "commonity" in frequency}

\item{tokenized_corpus}{(Optional) A tokenized version of the corpus used to find word frequencies}
}
\value{
A filtered corpus
}
\description{
Removes common words from corpus
}
