% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_emails}
\alias{rm_emails}
\title{Removes emails from corpus}
\usage{
rm_emails(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Removes emails from corpus
}
