% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_non_alphanumeric}
\alias{rm_non_alphanumeric}
\title{Removes non-alphanumeric characters from corpus}
\usage{
rm_non_alphanumeric(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Removes non-alphanumeric characters from corpus
}
