% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_numbers}
\alias{rm_numbers}
\title{Removes numbers from corpus}
\usage{
rm_numbers(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
It does not remove numbers embedded in words such as 'hello123', '123hello', 'h3ll0'.
}
