% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_rare_words}
\alias{rm_rare_words}
\title{Removes rare words from corpus}
\usage{
rm_rare_words(corpus, rare_word_limit, tokenized_corpus = NULL)
}
\arguments{
\item{corpus}{A dataframe of the corpus}

\item{rare_word_limit}{An integer defining the rarity in frequency}

\item{tokenized_corpus}{(Optional) A tokenized version of the corpus used to find word frequencies}
}
\value{
A filtered corpus
}
\description{
Removes rare words from corpus
}
