% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_regexp}
\alias{rm_regexp}
\title{Removes based on regular expression}
\usage{
rm_regexp(corpus, pattern)
}
\arguments{
\item{corpus}{A dataframe of the corpus}

\item{pattern}{The regular expression}
}
\value{
A filtered corpus
}
\description{
Removes based on regular expression
}
