% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_urls}
\alias{rm_urls}
\title{Removes urls from corpus}
\usage{
rm_urls(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Removes urls from corpus
}
