% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_whitespace}
\alias{rm_whitespace}
\title{Removes extra whitespace from corpus}
\usage{
rm_whitespace(corpus)
}
\arguments{
\item{corpus}{A dataframe of the corpus}
}
\value{
A filtered corpus
}
\description{
Removes extra whitespace from corpus
}
