% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{rm_words}
\alias{rm_words}
\title{Removes specified words from corpus}
\usage{
rm_words(corpus, words, tokenized_corpus = NULL)
}
\arguments{
\item{corpus}{A dataframe of the corpus}

\item{tokenized_corpus}{Not yet implemented}

\item{rare_word_limit}{An integer defining the rarity in frequency}
}
\value{
A filtered corpus
}
\description{
Removes specified words from corpus
}
