% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_lowercase}
\alias{tf_lowercase}
\title{Transforms corpus to lowercase}
\usage{
tf_lowercase(corpus)
}
\arguments{
\item{corpus}{Corpus to transform}
}
\value{
A transformed corpus
}
\description{
Transforms corpus to lowercase
}
