% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_merge_tokens}
\alias{tf_merge_tokens}
\title{Transforms tokens to corpus based on document id}
\usage{
tf_merge_tokens(tokenized_corpus, delim = " ")
}
\arguments{
\item{tokenized_corpus}{Tokenized corpus to merge}

\item{delim}{Delimiter between tokens}
}
\value{
A dataframe with each document as id and text columns
}
\description{
Transforms tokens to corpus based on document id
}
