% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_tokenize}
\alias{tf_tokenize}
\title{Transforms corpus to tokens}
\usage{
tf_tokenize(corpus, token = "words", ...)
}
\arguments{
\item{corpus}{Corpus to transform}

\item{token}{Definition of token (see \code{tidytext::unnest_tokens} for options)}
}
\value{
A dataframe with token column containing each token
}
\description{
Transforms corpus to tokens
}
